mcp-client-python/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "André", email = "andre.pereira66170@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.72.0",
    "mcp>=1.20.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
